feat(player): add hp and change double to float

+++ src/game/tile/i_box.hpp
@@ -8,8 +8,8 @@ namespace ep::game
   public:
     virtual ~IBox() = default;
 
-    virtual double GetX() const noexcept = 0;
-    virtual double GetY() const noexcept = 0;
+    virtual float GetX() const noexcept = 0;
+    virtual float GetY() const noexcept = 0;
     virtual std::uint8_t GetWidth() const noexcept = 0;
     virtual std::uint8_t GetHeight() const noexcept = 0;
   };
